menu and system fully working with shift and sub
